fixes: - deprecated 'cancelOnChange'. The 'abort' function should be used instead - added 'params' feature in the vanilla fetcher - added useful imports like 'useData', 'useLoading', 'useConfig', etc
